+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - CascadiaMono is no longer the default font
  - No longer using dock widgets, Instruments + Waveforms are now located to
    the right of the pattern editor.
+ - Pattern editor row numbers can be displayed in decimal (default is hexadecimal)
 ### Fixed
  - Issue with Instrument Editor not properly updating when changing instruments
  - Hitting enter in the sequence editor line edit opens the Wave editor